requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
babel==2.18.0
certifi==2026.2.25
charset-normalizer==3.4.5
colorama==0.4.6
courlan==1.3.2
dateparser==1.3.0
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
executing==2.2.1
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.13.0
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.4
matplotlib-inline==0.2.1
mysql-connector-python==9.6.0
numpy==2.4.3
openai==2.26.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
prompt_toolkit==3.0.52
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.6
pytz==2026.1.post1
regex==2026.2.28
requests==2.32.5
setuptools==80.10.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tld==0.13.2
tqdm==4.67.3
trafilatura==2.0.0
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
wcwidth==0.6.0
wheel==0.46.3
